# internal/platform/apidocs/swagger.yaml
definitions:
...
  apierr.Envelope:
    properties:
      error:
        $ref: '#/definitions/apierr.EnvelopeEr'
      request_id:
        example: req_example
        type: string
      type:
        enum:
        - error
        example: error
        type: string
    required:
    - error
    - request_id
    - type
    type: object
  apierr.EnvelopeEr:
    properties:
      details:
        additionalProperties: {}
        type: object
      message:
        example: input is required
        type: string
      type:
        enum:
        - invalid_request_error
        - authentication_error
        - permission_error
        - not_found_error
        - request_too_large
        - rate_limit_error
        - api_error
        - timeout_error
        - overloaded_error
        example: invalid_request_error
        type: string
    required:
    - message
    - type
    type: object
  deployments.CreateRequest:
    properties:
      agent_id:
        type: string
      cron:
        type: string
      environment_id:
        type: string
      input:
        type: string
      name:
        type: string
    required:
    - agent_id
    - input
    - name
    type: object
  deployments.Deployment:
    properties:
      agent_id:
        type: string
      created_at:
        format: date-time
        type: string
      cron:
        type: string
      environment_id:
        type: string
      id:
        type: string
      input:
        type: string
      last_task_id:
        type: string
      name:
        type: string
      next_at:
        format: date-time
        type: string
      paused:
        type: boolean
    required:
    - agent_id
    - created_at
    - id
    - input
    - name
    - paused
    type: object
  deployments.ListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/deployments.Deployment'
        type: array
      next_offset:
        description: 仅有下一页时返回；省略表示末页。 || Returned only when another page exists; omitted
          on the last page.
        type: integer
    required:
    - data
    type: object
  deployments.PauseRequest:
    properties:
      paused:
        type: boolean
        x-nullable: true
    required:
    - paused
    type: object
  deployments.Run:
    properties:
      created_at:
        format: date-time
        type: string
      deployment_id:
        type: string
      id:
        type: string
      reason:
        type: string
      session_id:
        type: string
      task_id:
        type: string
    required:
    - created_at
    - deployment_id
    - id
    type: object
  deployments.RunsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/deployments.Run'
        type: array
      next_offset:
        description: 仅有下一页时返回；省略表示末页。 || Returned only when another page exists; omitted
          on the last page.
        type: integer
    required:
    - data
    type: object
...
info:
  contact: {}
  description: |-
    Managed Agents 服务。所有 /v1 接口使用 Bearer Wave API Key 认证； || Managed Agents service. All /v1 endpoints require a Bearer Wave API Key;
    由 `wave bootstrap` 签发。Authorization 填写完整的 `Bearer <Wave API Key>`。 || issue it with `wave bootstrap`. Set Authorization to the full `Bearer <Wave API Key>` value.
    offset 分页返回 data 和可选 next_offset（省略表示末页）；消息历史使用 next_after，事件列表使用最后一条 sequence 续读。 || Offset pagination returns data and an optional next_offset (omitted on the last page); message history uses next_after, and event lists resume from the last sequence.
    JSON 错误返回 {"type":"error","error":{"type":"api_error","message":"..."},"request_id":"req_..."}。文件传输的 Range 错误返回 text/plain；条件下载和健康检查可返回空响应。 || JSON errors return {"type":"error","error":{"type":"api_error","message":"..."},"request_id":"req_..."}. File-transfer Range errors return text/plain; conditional downloads and health checks may return empty responses.
    JSON 请求拒绝未知字段，非 multipart 请求体上限为 2 MiB。可空字段通过 x-nullable 标记；省略与 null 不等价。 || JSON requests reject unknown fields; non-multipart request bodies are limited to 2 MiB. Nullable fields use x-nullable; omission is not equivalent to null.
    交互式文档见 /swagger/index.html，OpenAPI spec 见 /swagger/doc.json。 || Interactive documentation: /swagger/index.html. OpenAPI spec: /swagger/doc.json.
  title: Wave AI API
  version: "1.0"
paths:
...
  /v1/deployments:
    get:
      description: 分页返回当前所有者可见的部署。 || Return deployments visible to the current owner,
        with pagination.
      operationId: deploymentsList
      parameters:
      - default: 100
        description: 每页条数 || Items per page
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: 偏移量 || Offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/deployments.ListResponse'
        "400":
          description: 分页参数不合法 || Invalid pagination parameters
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 列出部署 || List deployments
      tags:
      - deployments
    post:
      consumes:
      - application/json
      description: 创建一个将 Agent 与可选环境绑定、按输入运行的部署；cron 为标准五段表达式，填写时计算下次触发时间，为空则仅手动触发。
        || Create a deployment that binds an Agent to an optional environment and
        runs the supplied input. cron uses the standard five-field syntax; when set,
        the next run is calculated, otherwise runs are manual only.
      operationId: deploymentsCreate
      parameters:
      - description: 部署名称、Agent ID、可选环境 ID、输入与 cron || Deployment name, Agent ID,
          optional environment ID, input and cron
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/deployments.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/deployments.Deployment'
        "400":
          description: 请求体不是合法 JSON 或 cron 表达式不合法 || Invalid JSON body or cron expression
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "404":
          description: Agent 或环境不存在 || Agent or environment not found
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 创建部署 || Create a deployment
      tags:
      - deployments
  /v1/deployments/{id}:
    patch:
      consumes:
      - application/json
      description: 设置部署的暂停状态；恢复时会重新计算 cron 的下次触发时间。 || Set the deployment's paused
        state; resuming recalculates the next cron run.
      operationId: deploymentsPause
      parameters:
      - description: 部署 ID || Deployment ID
        in: path
        name: id
        required: true
        type: string
      - description: 目标暂停状态 || Desired paused state
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/deployments.PauseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/deployments.Deployment'
        "400":
          description: 请求体不是合法 JSON 或缺少 paused 字段 || Invalid JSON body or missing
            paused field
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "404":
          description: 部署不存在或不在当前所有者范围内 || Deployment not found or outside the current
            owner's scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 暂停或恢复部署 || Pause or resume a deployment
      tags:
      - deployments
  /v1/deployments/{id}/run:
    post:
      description: 立即触发一次部署运行，创建会话与任务并异步执行；部署处于暂停状态时返回 409。 || Trigger a deployment
        immediately, creating a session and task for asynchronous execution. A paused
        deployment returns 409.
      operationId: deploymentsRun
      parameters:
      - description: 部署 ID || Deployment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/deployments.Run'
        "400":
          description: 环境或 Agent 已归档，或任务资源校验失败 || Environment or Agent is archived,
            or task resource validation failed
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "404":
          description: 部署不存在或不在当前所有者范围内 || Deployment not found or outside the current
            owner's scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "409":
          description: 部署已暂停 || Deployment is paused
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 手动触发部署 || Trigger a deployment manually
      tags:
      - deployments
  /v1/deployments/{id}/runs:
    get:
      description: 分页返回该部署的触发记录，包含关联的任务与会话 ID。 || Return run records for the deployment,
        with pagination, including the associated task and session IDs.
      operationId: deploymentsRuns
      parameters:
      - description: 部署 ID || Deployment ID
        in: path
        name: id
        required: true
        type: string
      - default: 100
        description: 每页条数 || Items per page
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: 偏移量 || Offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/deployments.RunsResponse'
        "400":
          description: 分页参数不合法 || Invalid pagination parameters
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "404":
          description: 部署不存在或不在当前所有者范围内 || Deployment not found or outside the current
            owner's scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 部署运行记录 || Deployment run records
      tags:
      - deployments
...
securityDefinitions:
  BearerAuth:
    description: 填写完整的 Bearer <Wave API Key>，包含 Bearer 前缀和空格。 || Enter the full Bearer
      <Wave API Key> value, including the Bearer prefix and a space.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"